M0078DDG-368 Make warnings fail the build
This change updates the sphinx and readthedocs configuration to make the build fails if there are warnings.
The change done in the Makefile allows to apply this change when building the doc locally.
The change done in the file `.readthedocs.yml` allows to apply this change when building the doc on ReadTheDocs.