M0137W-616 Enable linkchecker and remove failing link
This change enables the linkchecker and removes the failing link : https://www.java.com/en/download/manual.jsp
This link seems correct but for some reasons it does not pass the linkchecker.

Also the linkchecker timeout has been set to 20 seconds to avoid to be stuck too much time on a failing link.